+++ src/Models/AgentProfile.php
@@ -2,6 +2,7 @@
 
 namespace Escalated\Laravel\Models;
 
+use Escalated\Laravel\Concerns\UsesEscalatedConnection;
 use Escalated\Laravel\Enums\ChatStatus;
 use Escalated\Laravel\Escalated;
 use Illuminate\Database\Eloquent\Model;
@@ -9,6 +10,8 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 class AgentProfile extends Model
 {
+    use UsesEscalatedConnection;
+
     protected $guarded = ['id'];
 
     public function getTable(): string