+++ src/Models/SatisfactionRating.php
@@ -2,6 +2,7 @@
 
 namespace Escalated\Laravel\Models;
 
+use Escalated\Laravel\Concerns\UsesEscalatedConnection;
 use Escalated\Laravel\Escalated;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -9,6 +10,8 @@ use Illuminate\Database\Eloquent\Relations\MorphTo;
 
 class SatisfactionRating extends Model
 {
+    use UsesEscalatedConnection;
+
     public $timestamps = false;
 
     protected $guarded = ['id'];